move dropdown listitem to DdListMapper
